+++ equinix/provider.go
@@ -3,7 +3,6 @@ package equinix
 import (
 	"context"
 	"fmt"
-	"reflect"
 	"strings"
 	"time"
 
@@ -23,15 +22,6 @@ var (
 	NetworkTypeList      = strings.Join(DeviceNetworkTypes, ", ")
 	NetworkTypeListHB    = strings.Join(DeviceNetworkTypesHB, ", ")
 )
-
-// resourceDataProvider provies interface to schema.ResourceData
-// for convenient mocking purposes
-type resourceDataProvider interface {
-	Get(key string) interface{}
-	GetOk(key string) (interface{}, bool)
-	HasChange(key string) bool
-	GetChange(key string) (interface{}, interface{})
-}
 
 // Provider returns Equinix terraform *schema.Provider
 func Provider() *schema.Provider {
@@ -249,40 +239,6 @@ func isStringInSlice(needle string, hay []string) bool {
 	return false
 }
 
-func getResourceDataChangedKeys(keys []string, d resourceDataProvider) map[string]interface{} {
-	changed := make(map[string]interface{})
-	for _, key := range keys {
-		if v := d.Get(key); v != nil && d.HasChange(key) {
-			changed[key] = v
-		}
-	}
-	return changed
-}
-
-func getResourceDataListElementChanges(keys []string, listKeyName string, listIndex int, d resourceDataProvider) map[string]interface{} {
-	changed := make(map[string]interface{})
-	if !d.HasChange(listKeyName) {
-		return changed
-	}
-	old, new := d.GetChange(listKeyName)
-	oldList := old.([]interface{})
-	newList := new.([]interface{})
-	if len(oldList) < listIndex || len(newList) < listIndex {
-		return changed
-	}
-	return getMapChangedKeys(keys, oldList[listIndex].(map[string]interface{}), newList[listIndex].(map[string]interface{}))
-}
-
-func getMapChangedKeys(keys []string, old, new map[string]interface{}) map[string]interface{} {
-	changed := make(map[string]interface{})
-	for _, key := range keys {
-		if !reflect.DeepEqual(old[key], new[key]) {
-			changed[key] = new[key]
-		}
-	}
-	return changed
-}
-
 func isEmpty(v interface{}) bool {
 	switch v := v.(type) {
 	case int:
@@ -346,15 +302,4 @@ func slicesMatchCaseInsensitive(s1, s2 []string) bool {
 		}
 	}
 	return true
-}
-
-func schemaSetToMap(set *schema.Set) map[int]interface{} {
-	transformed := make(map[int]interface{})
-	if set != nil {
-		list := set.List()
-		for i := range list {
-			transformed[set.F(list[i])] = list[i]
-		}
-	}
-	return transformed
 }
