+++ equinix/provider_test.go
@@ -3,15 +3,12 @@ package equinix
 import (
 	"fmt"
 	"os"
-	"reflect"
 	"regexp"
 	"strings"
 	"testing"
 
-	"github.com/equinix/terraform-provider-equinix/internal/config"
-	"github.com/equinix/terraform-provider-equinix/internal/hashcode"
-
 	"github.com/equinix/ecx-go/v2"
+	"github.com/equinix/terraform-provider-equinix/internal/config"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/stretchr/testify/assert"
@@ -23,28 +20,6 @@ var (
 	testAccProvider          *schema.Provider
 	testExternalProviders    map[string]resource.ExternalProvider
 )
-
-type mockedResourceDataProvider struct {
-	actual map[string]interface{}
-	old    map[string]interface{}
-}
-
-func (r mockedResourceDataProvider) Get(key string) interface{} {
-	return r.actual[key]
-}
-
-func (r mockedResourceDataProvider) GetOk(key string) (interface{}, bool) {
-	v, ok := r.actual[key]
-	return v, ok
-}
-
-func (r mockedResourceDataProvider) HasChange(key string) bool {
-	return !reflect.DeepEqual(r.old[key], r.actual[key])
-}
-
-func (r mockedResourceDataProvider) GetChange(key string) (interface{}, interface{}) {
-	return r.old[key], r.actual[key]
-}
 
 type mockECXClient struct {
 	GetUserPortsFn func() ([]ecx.Port, error)
@@ -176,82 +151,6 @@ func TestProvider_stringsFound_negative(t *testing.T) {
 	assert.False(t, result, "Given strings were found")
 }
 
-func TestProvider_resourceDataChangedKeys(t *testing.T) {
-	// given
-	keys := []string{"key", "keyTwo", "keyThree"}
-	rd := mockedResourceDataProvider{
-		actual: map[string]interface{}{
-			"key":    "value",
-			"keyTwo": "newValueTwo",
-		},
-		old: map[string]interface{}{
-			"key":    "value",
-			"keyTwo": "valueTwo",
-		},
-	}
-	expected := map[string]interface{}{
-		"keyTwo": "newValueTwo",
-	}
-	// when
-	result := getResourceDataChangedKeys(keys, rd)
-	// then
-	assert.Equal(t, expected, result, "Function returns valid key changes")
-}
-
-func TestProvider_resourceDataListElementChanges(t *testing.T) {
-	// given
-	keys := []string{"key", "keyTwo", "keyThree"}
-	listKeyName := "myList"
-	rd := mockedResourceDataProvider{
-		old: map[string]interface{}{
-			listKeyName: []interface{}{
-				map[string]interface{}{
-					"key":      "value",
-					"keyTwo":   "valueTwo",
-					"keyThree": 50,
-				},
-			},
-		},
-		actual: map[string]interface{}{
-			listKeyName: []interface{}{
-				map[string]interface{}{
-					"key":      "value",
-					"keyTwo":   "newValueTwo",
-					"keyThree": 100,
-				},
-			},
-		},
-	}
-	expected := map[string]interface{}{
-		"keyTwo":   "newValueTwo",
-		"keyThree": 100,
-	}
-	// when
-	result := getResourceDataListElementChanges(keys, listKeyName, 0, rd)
-	// then
-	assert.Equal(t, expected, result, "Function returns valid key changes")
-}
-
-func TestProvider_mapChanges(t *testing.T) {
-	// given
-	keys := []string{"key", "keyTwo", "keyThree"}
-	old := map[string]interface{}{
-		"key":    "value",
-		"keyTwo": "valueTwo",
-	}
-	new := map[string]interface{}{
-		"key":    "newValue",
-		"keyTwo": "valueTwo",
-	}
-	expected := map[string]interface{}{
-		"key": "newValue",
-	}
-	// when
-	result := getMapChangedKeys(keys, old, new)
-	// then
-	assert.Equal(t, expected, result, "Function returns valid key changes")
-}
-
 func TestProvider_isEmpty(t *testing.T) {
 	// given
 	input := []interface{}{
@@ -329,31 +228,6 @@ func TestProvider_slicesMatch(t *testing.T) {
 	for i := range expected {
 		assert.Equal(t, expected[i], results[i])
 	}
-}
-
-func TestProvider_schemaSetToMap(t *testing.T) {
-	// given
-	type item struct {
-		id       string
-		valueOne int
-		valueTwo int
-	}
-	setFunc := func(v interface{}) int {
-		i := v.(item)
-		return hashcode.String(i.id)
-	}
-	items := []interface{}{
-		item{"id1", 100, 200},
-		item{"id2", 666, 999},
-		item{"id3", 0, 100},
-	}
-	set := schema.NewSet(setFunc, items)
-	// when
-	list := schemaSetToMap(set)
-	// then
-	assert.Equal(t, items[0], list[setFunc(items[0])])
-	assert.Equal(t, items[1], list[setFunc(items[1])])
-	assert.Equal(t, items[2], list[setFunc(items[2])])
 }
 
 //‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾
