+++ internal/schema/resourcedata.go
@@ -1,0 +1,46 @@
+package schema
+
+import "reflect"
+
+// resourceDataProvider provies interface to schema.ResourceData
+// for convenient mocking purposes
+type resourceDataProvider interface {
+	Get(key string) interface{}
+	GetOk(key string) (interface{}, bool)
+	HasChange(key string) bool
+	GetChange(key string) (interface{}, interface{})
+}
+
+func getMapChangedKeys(keys []string, old, new map[string]interface{}) map[string]interface{} {
+	changed := make(map[string]interface{})
+	for _, key := range keys {
+		if !reflect.DeepEqual(old[key], new[key]) {
+			changed[key] = new[key]
+		}
+	}
+	return changed
+}
+
+func GetResourceDataChangedKeys(keys []string, d resourceDataProvider) map[string]interface{} {
+	changed := make(map[string]interface{})
+	for _, key := range keys {
+		if v := d.Get(key); v != nil && d.HasChange(key) {
+			changed[key] = v
+		}
+	}
+	return changed
+}
+
+func GetResourceDataListElementChanges(keys []string, listKeyName string, listIndex int, d resourceDataProvider) map[string]interface{} {
+	changed := make(map[string]interface{})
+	if !d.HasChange(listKeyName) {
+		return changed
+	}
+	old, new := d.GetChange(listKeyName)
+	oldList := old.([]interface{})
+	newList := new.([]interface{})
+	if len(oldList) < listIndex || len(newList) < listIndex {
+		return changed
+	}
+	return getMapChangedKeys(keys, oldList[listIndex].(map[string]interface{}), newList[listIndex].(map[string]interface{}))
+}
