+++ internal/schema/set_map.go
@@ -1,4 +1,4 @@
-package utils
+package schema
 
 import (
 	"sort"
